Allow get options to be passed to .get()
This would allow us to specify `{ source: cache}` in `fetch` and `fetchAndAdd` actions to read directly from cache instead of from the server. This could potentially save a lot of reads in the case where we know there is no new content and would prefer to read from cache directly if possible.

Thanks for an awesome library!